Add a mechanism for re-creating a runner.
When a pot for a runner is first created, the runner configuration is
copied outside of the pot / jail.  This can now be used to create a new
pot with the same configuration.

This is not an ideal solution as it depends on implementation details of
pot.

This, combined with the user-provided flavours support, makes it
possible to automatically update runners to the latest version of all
installed software.

Fixes #1