Rework to use clones instead of snapshots.
Provide an atomic update mechanism for the base image (we can replace it
while the clone is running, we just need to block clone creation while
we are doing the destroy + rename).

Add a wrapper script that creates all runners.  This can be run from a
cron job to periodically recreate all runners and pick up all updates
and so on.